requirements.txt
Pillow==10.4.0
requests==2.32.5
openai==1.44.0
blend-modes==2.1.0
huggingface_hub>=0.34.0
color_matcher==0.5.0
chardet==5.2.0
google-generativeai==0.7.2
anthropic